+++ src/main/java/POJOs/GameData.java
@@ -1,13 +1,13 @@
 package POJOs;
 
 public class GameData {
-    private int id;
-    private int points;
+    private long id;
+    private long points;
     private String gameId;
     private String createdAt;
     private Player player1, player2;
 
-    public GameData(int id, int points, String gameId, String createdAt, Player player1, Player player2) {
+    public GameData(long id, long points, String gameId, String createdAt, Player player1, Player player2) {
         this.id = id;
         this.points = points;
         this.gameId = gameId;
@@ -16,11 +16,11 @@ public class GameData {
         this.player2 = player2;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public int getPoints() {
+    public long getPoints() {
         return points;
     }
